Implement Filament Loading
Implemented filament loading scripts for a variaty of filaments. Duet doesnt have variables yet so it uses an extra tool to store the filament name, temprature and extrude speed for the loading macro to use.

Setting the extrude speed is busted in 3.1.1, see https://forum.duet3d.com/topic/18802/how-to-g1-f-expression/8

Set the retract length and speed to enable machine based retracts in the future

Also renamed tools so they start at "T1" in the GUI instead of "T0"