Refactor from "Virtual" Axes to Work Offsets
Refactor all of the code using the virtual axes to use the work coordinate system instead. This requires a script to assign variables when a tool number I known as it requires 8 line of code to do. Totally untested.

Also I switched to 0.9 degree steppers on the X and Y axes. This commit has the steps/mm changes.