Switch based homing
Use switches for homing instead of stall detection

Home X requires the Y axis to be in a specific location. Homing when it is not in that location would result in a crash. X homing validates that Y is homed and moves Y with a stall detect move before it proceeds to home X.

Call M501 in globals to restore machine state before overwriting with configuration.

Updates to the motor settings for the 0.9 degree steppers. The ones I am testing with are a little small for the application. Raised the speed where the switch to spread cycle so tool changes happen in spread cycle mode.