tart of the tool changing re-write
Initial pass at a more optimized tool change procedure. The primary goal is to use the tools individual Y offset, the length of the dock pins, and the length of the T-pin to make moves that are safe for a variety of tools of different depths.

No testing yet as the achine is down for a PSU issue.