Tweaks to Tool Changing
this sets the safe distances to be closer to save on meovement time. This has been tested and looks to be a very gentle and reliable tool change.

Also tweaked the order of baby stepping restore so it is reflected in the UI before the program blocks for the heater.