Irregular Mesh Bed Probing
Add the two scripts required to probe irregular points on the bed and produce a heightmap that Duet can use.

A Python script generates the GCode for probing using a set of points embedded in the script. The points are reflected for bed symetry and repeating rows are expanded so its quicker to make updates if needed.

The probing macro takes advantage of the Debug logging facility on the Duet to capture the probed Z values from G30 calls. This log is then processed by a Python program to produce the height map. Eventuall This program will be triggered synchronously by a custom GCode through https://github.com/wilriker/execonmcode

The height map processor uses Radial Basis Functions imported from SciPy to create a surface from the probed points. This surface is then sampled in a regular pattern at the locations required for a heightmap.csv file that RRF can work with.

RBF allows sampling beyond the convex hull of the probed points which helps get the heightmap mesh to cover the entire print area, even though the inductive probe has to stay at least 5mm inside the boundary to get consistent readings.

I've had several successful test run and experimented with multiple variations of the GCode. I tried versions with sub-macro calls and lots of usage of the object model and expression language. This performed worse for me than plain GCode. This commit generates plain GCode with no use of expressions or the Object model. If I can resolve the performance issues I might go back to expressions.