Updated mesh probing for improves SPI communication
Based on this investigation: https://forum.duet3d.com/topic/21077/slow-unreliable-spi-communication-when-logging-with-m929
I swapped my SD card and upped the SPI transmission frequency, polling time and buffer sizes. This resolved all of the slowdowns I was seeing in previous tests so I have gone back to the inital version that calls a macro to probe each point and uses object model variables for speeds and offsets. This makes the macro more flexible.