Perform StealthChop auto tuning while homing
This adds a much needed script to tune StealthChop. This was a huge improvement in noise profile and likely is the root cause of a lot of noise issues. Unfortunatly it screeches terrible above 180mm/sec so the top speed of the printer had to be limited to 150mm/sec.