Try to fix C homing reliability
Failed C homing has been a recurring issue. The axis does not like to move at reduced power. It also never actually detects a stall.

This change "kicks" the axis into motion at full power and then drops the power to 65% for the rest of the homing move. The move was shortened to reduce the time the stepper is being driven into the hard stop. In config.g the stall detectio has been disabled.