Add acceleration limits to tool changing and homing
As I'm messing with accelerations and jerk settings it became apparent that the first few moves the printer does are not acceleration constrained. This is a big miss on my part. This adds acceleration caps to all homing and tool changing tasks. Maybe later these can be enhanced to consider if the gantry is loaded or unloaded.